+++ cmd/template_nomad.go
@@ -29,24 +29,30 @@ var templateNomadCmd = &cobra.Command{
 
 		return templateNomad(string(template), networkState)
 	},
+	Example: `
+# Generate the nomad configuration for multiple node sets #1
+vegacapsule template nomad --path .../node_set.tmpl --nodeset-group-name validators,full
+
+# Generate the nomad configuration for multiple node sets #2
+vegacapsule template nomad --path .../node_set.tmpl --nodeset-group-name validators --nodeset-group-name full`,
 }
 
 func init() {
-	templateNomadCmd.PersistentFlags().StringVar(&nodeSetGroupName,
+	templateNomadCmd.PersistentFlags().StringSliceVar(&nodeSetsGroupsNames,
 		"nodeset-group-name",
-		"",
+		[]string{},
 		"Allows to apply template to all node sets in a specific group",
 	)
 
-	templateNomadCmd.PersistentFlags().StringVar(&nodeSetName,
+	templateNomadCmd.PersistentFlags().StringSliceVar(&nodeSetsNames,
 		"nodeset-name",
-		"",
+		[]string{},
 		"Allows to apply template to a specific node set",
 	)
 }
 
 func templateNomad(templateRaw string, netState *state.NetworkState) error {
-	nodeSets, err := getNodeSetsByNames(netState)
+	nodeSets, err := filterNodesSets(netState, nodeSetsNames, nodeSetsGroupsNames)
 	if err != nil {
 		return err
 	}
@@ -58,7 +64,7 @@ func templateNomad(templateRaw string, netState *state.NetworkState) error {
 		}
 
 		fileName := fmt.Sprintf("nomad-%s.hcl", ns.Name)
-		if err := outputTemplate(buff, fileName); err != nil {
+		if err := outputTemplate(buff, templateOutDir, fileName, true); err != nil {
 			return err
 		}
 	}